Move links module into facets package
